refactor: checkProductPurchaseMenu.js에서 동전 반환 관련 파일 분리
- CheckCoinReturn.js 파일에서 동전 반환 관련 로직 담당
